python-service/requirements.txt
#require flask framework
Flask==1.0.2

#require dotenv
python-dotenv

#require google cloud vision library
google-cloud-vision